Change `inspector` to `configuration` for clarity
This makes it more clear what the commands do.
