+++ api/configuration/config.go
@@ -1,4 +1,4 @@
-package inspector
+package configuration
 
 type Config struct {
 	GitHubActionVersionPinningCheckEnabled bool `yaml:"github_action_version_pinning_check_enabled"`
