+++ kusari/cmd/configuration.go
@@ -1,0 +1,15 @@
+package cmd
+
+import "github.com/spf13/cobra"
+
+func KusariConfiguration() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "configuration",
+		Short: "Configuration",
+		Long:  "Configuration actions",
+	}
+
+	cmd.AddCommand(generateConfig())
+
+	return cmd
+}
